pyproject.toml

[tool.poetry]
name = "julia_frost_ts"
version = "1.0.0"
description = "weather data to time series data in CDF"
authors = ["Juliamg <julia.graham@live.no>"]

[tool.black]
line-length = 120
target_version = ['py37']
include = '\.py$'

[tool.isort]
line_length=120                # corresponds to -w  flag
multi_line_output=3            # corresponds to -m  flag
include_trailing_comma=true    # corresponds to -tc flag
skip_glob = '^((?!py$).)*$'    # this makes sort all Python files
known_third_party = ["cognite", "dotenv", "requests"]

[tool.poetry.dependencies]
python = "^3.7"
cognite-sdk = "^2.38.3"
cognite-extractor-utils = "^1.5.4"
python-dotenv = "^0.19.2"

[tool.poetry.dev-dependencies]
pyinstaller = "^4.0"
macholib = {version = "^1.14", platform = "darwin"}             # Used by pyinstaller pn Mac OS
pywin32-ctypes = {version = "^0.2.0", platform = "win32"}       # Used by pyinstaller on Windows
pefile = "^2019.4.18"                                           # Used by pyinstaller on Windows
mypy = "^0.920"
flake8 = "^4.0.1"
black = "^21.12b0"
isort = "^5.10.1"
pre-commit = "^2.16.0"

[tool.poetry.scripts]
julia_frost_ts = "julia_frost_ts.__main__:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
